[FIXED] Fixed a crash caused by unwrapping an Optional value while running ACME Swift on iPad.
### Steps to reproduce

Launch ACME Swift (`App Demo for iOS Swift`) on an iPad.

#### Expected

The app should simply run, even if no images are available in `Images.xcassets`. This should and will be addressed separately

#### Actual

Bam! It crashes on launch with the `"fatal error: unexpectedly found nil while unwrapping an Optional value"` error message in the console.